bedu-django-ci/requirements.txt
Django==3.1.3
psycopg2==2.8.6
dj-database-url==0.5.0
djangorestframework==3.12.2

django-extensions==3.0.9

coverage==5.3
selenium==3.141.0
flake8==3.8.4
mypy==0.790
